Move responsibility for outputting barlines to BarManager
This delays printing barline until something happens with a duration or
at the end of the piece. It allows us to ensure we output the correct
barline when there is a repeat and should make it very easy to support
custom barlines.

(The only downside is that currently it will only print the barline at
an actual end of a measure, but that shouldn't be too difficult to
change if needed)